requirements.txt
fastapi==0.121.2
uvicorn[standard]==0.38.0
sqlalchemy==2.0.44
psycopg2-binary==2.9.11
alembic==1.17.2
pydantic==2.12.4
python-dotenv==1.2.1
twilio==9.8.6
python-multipart==0.0.20
